backend/requirements.txt
Flask==2.3.2
Flask-PyMongo==2.3.0
bcrypt==3.2.0
PyJWT==2.9.0
flask-cors=5.0.0